+++ fabric/src/main/java/net/anvian/chiseledbookshelfvisualizer/client/config/ModMenuIntegration.java
@@ -1,0 +1,13 @@
+package net.anvian.chiseledbookshelfvisualizer.client.config;
+
+import com.terraformersmc.modmenu.api.ConfigScreenFactory;
+import com.terraformersmc.modmenu.api.ModMenuApi;
+import me.shedaniel.autoconfig.AutoConfigClient;
+
+public class ModMenuIntegration implements ModMenuApi {
+    @Override
+    public ConfigScreenFactory<?> getModConfigScreenFactory() {
+        return parent ->
+                AutoConfigClient.getConfigScreen(ClientConfig.class, parent).get();
+    }
+}